+++ android/app/src/main/AndroidManifest.xml
@@ -31,6 +31,11 @@
         </activity>
 
         <activity
+            android:name="com.jaino.setting.account.AccountActivity"
+            android:exported="true">
+        </activity>
+
+        <activity
             android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
             android:exported="true">
             <intent-filter>
