--- android/feature/setting/src/main/java/com/jaino/setting/account/AccountFragment.kt
@@ -1,56 +1,0 @@
-package com.jaino.setting.account
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
-import com.jaino.setting.R
-import com.jaino.setting.databinding.FragmentAccountBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class AccountFragment : Fragment() {
-
-    private val viewModel : AccountViewModel by viewModels()
-
-    private var _binding: FragmentAccountBinding? = null
-    private val binding
-        get() = requireNotNull(_binding) { "binding object is not initialized" }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_account, container, false)
-        binding.lifecycleOwner = viewLifecycleOwner
-        binding.viewModel = viewModel
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        initButtons()
-    }
-
-    private fun initButtons(){
-        binding.accountBackButton.setOnClickListener {
-            navigateToSetting()
-        }
-    }
-
-    private fun navigateToSetting(){
-        findNavController().navigate(
-            R.id.action_accountFragment_to_settingFragment
-        )
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
